requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.32
alembic>=1.13.3
pydantic>=2.9.2
pydantic-settings>=2.5.2
pymysql>=1.1.1
python-multipart>=0.0.9
